+++ core/src/indexer/start.rs
@@ -1,17 +1,8 @@
-use std::{collections::HashMap, path::Path, str::FromStr, sync::Arc, time::Duration};
+use std::{collections::HashMap, path::Path, sync::Arc, time::Duration};
 
-use colored::Colorize;
-use ethers::{
-    abi::Hash,
-    prelude::ActionType,
-    providers::ProviderError,
-    types::{Action, Address, Bytes, U256, U64},
-};
+use ethers::{providers::ProviderError, types::U64};
 use futures::future::try_join_all;
-use serde::Serialize;
-use serde_json::json;
 use tokio::{
-    join,
     sync::Semaphore,
     task::{JoinError, JoinHandle},
     time::{sleep, Instant},
@@ -21,14 +12,13 @@ use tracing::{error, info};
 use crate::{
     database::postgres::client::PostgresConnectionError,
     event::{
-        callback_registry::{EventCallbackRegistry, TxInformation},
-        config::EventProcessingConfig,
+        callback_registry::EventCallbackRegistry, config::EventProcessingConfig,
         contract_setup::NetworkContract,
-        EventMessage,
     },
     indexer::{
         dependency::ContractEventsDependenciesConfig,
         last_synced::{get_last_synced_block_number, SyncConfig},
+        native_transfer::native_transfer_block_consumer,
         process::{
             process_contracts_events_with_dependencies, process_event,
             ProcessContractsEventsWithDependenciesError, ProcessEventError,
@@ -38,7 +28,7 @@ use crate::{
         ContractEventDependencies,
     },
     manifest::core::Manifest,
-    provider::{CreateNetworkProvider, JsonRpcCachedProvider},
+    provider::CreateNetworkProvider,
     streams::StreamsClients,
     PostgresClient,
 };
@@ -90,119 +80,6 @@ pub enum StartIndexingError {
 pub struct ProcessedNetworkContract {
     pub id: String,
     pub processed_up_to: U64,
-}
-
-#[derive(Serialize)]
-pub struct NativeTransfer {
-    pub from: Address,
-    pub to: Address,
-    pub value: U256,
-    pub transaction_information: TxInformation,
-}
-
-async fn consumer_logic(
-    provider: Arc<JsonRpcCachedProvider>,
-    block_numbers: &[U64],
-    network_name: &str,
-    streams_client: Option<&StreamsClients>,
-) -> Result<(), StartIndexingError> {
-    let trace_futures: Vec<_> =
-        block_numbers.into_iter().map(|n| provider.trace_block(*n)).collect();
-
-    let trace_calls = try_join_all(trace_futures).await?;
-
-    let native_transfers = trace_calls
-        .into_iter()
-        .flatten()
-        .filter_map(|trace| {
-            let action = match trace.action {
-                Action::Call(call) => Some(call),
-                _ => None,
-            }?;
-
-            let has_value = !action.value.is_zero();
-            let no_input = action.input == Bytes::from_str("0x").unwrap();
-            let is_native_transfer = has_value && no_input;
-
-            if is_native_transfer {
-                if trace.transaction_hash.is_none() {
-                    error!("Transaction hash should exist for block trace {}", trace.block_number);
-                    return None;
-                }
-
-                Some(NativeTransfer {
-                    from: action.from,
-                    to: action.to,
-                    value: action.value,
-                    transaction_information: TxInformation {
-                        network: network_name.to_owned(),
-                        address: Address::zero(),
-                        block_number: U64::from(trace.block_number),
-                        block_timestamp: None,
-                        transaction_hash: trace.transaction_hash.expect("checked prior"),
-                        block_hash: trace.block_hash,
-                        transaction_index: U64::from(trace.transaction_position.unwrap_or(0)),
-                        log_index: U256::from(0),
-                    },
-                })
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-
-    if native_transfers.is_empty() {
-        return Ok(());
-    }
-
-    if let Some(client) = streams_client {
-        let contract_name = "EvmDebugTrace";
-        let event_name = "NativeTokenTransfer";
-        let from_block = native_transfers
-            .first()
-            .map(|b| b.transaction_information.block_number)
-            .unwrap_or_default();
-        let to_block = native_transfers
-            .last()
-            .map(|b| b.transaction_information.block_number)
-            .unwrap_or_default();
-
-        let stream_id = format!(
-            "{}-{}-{}-{}-{}",
-            contract_name, event_name, network_name, from_block, to_block
-        );
-
-        let event_message = EventMessage {
-            event_name: event_name.to_string(),
-            event_data: json!(native_transfers),
-            event_signature_hash: Hash::zero(),
-            network: network_name.to_string(),
-        };
-
-        match client.stream(stream_id, &event_message, false, true).await {
-            Ok(streamed) => {
-                if streamed > 0 {
-                    info!(
-                        "{}::{} - {} - {} events {}",
-                        contract_name,
-                        event_name,
-                        "STREAMED".green(),
-                        streamed,
-                        format!(
-                            "- trace block: {} - {} - network: {}",
-                            from_block, to_block, network_name
-                        )
-                    );
-                }
-            }
-            Err(e) => {
-                error!("Error streaming event: {}", e);
-                return Err(StartIndexingError::UnknownError(e.to_string()));
-            }
-        }
-    }
-
-    Ok(())
 }
 
 pub async fn start_indexing(
@@ -341,7 +218,7 @@ pub async fn start_indexing(
                         continue;
                     }
 
-                    let processed_block = consumer_logic(
+                    let processed_block = native_transfer_block_consumer(
                         provider.clone(),
                         &buffer[..recv],
                         &network_name,
